[#367] refactor: 인증남기기 버튼 색상 원복
